requirements.txt
discord                            1.7.3
python-dotenv                      0.19.0
textblob                           0.15.3